fixed comments and new deed selection test
Almost! :pensive: Unfortunately, you should assume that all tests are independent, so we can not expect that Feed # 0 will be loaded when we start this test, you also have to load it before. Moreover, remember to call the function to load the Feed #1 on the callback of the loading of Feed #0. For example:

loadFeed(0, function(){
      //do something
     loadFeed(1, function(){
          //do something
     });
});